Rename Project2-932810401.asm to Fibonacci Sequence - CS 271
Code written in MASM assembly code that does the follow: this programme asks the users their name, greets the user, introduces the programme, and then prompts the users for a whole number between 1-46. The programme then displays the Fibonacci sequence up to the number inputted by the user. The numbers are also displayed in rainbow colours as a cool trick for extra-credit.